Addition of Join Command
Without the join command, one would have to run the yt command twice, once to have the bot join you in a voice channel, and a second time to play the intended song. That simply won't do. (We have the yt command call the join command on not being in a voice channel anyway as before)