Feat/unique channel search (#2)
* Initial Attempt

The initial, untested attempt at the implementation of the bot's intended function: searching of a specific channel's video log. I've made it such that, if desired, one could use their own api key/channel id to use either their own resources or a youtube channel of their choosing, respectively.

* Search Implemented

After a great deal of toil in testing, the unique channel search works! Rather than using a direct filepath like I usually would when accesssing files, I found a solution that gets files out of a resources folder, provided it's in the right place. So, that's what is at play in fetching the properties. Going to implement that for my discord key as well, to (in theory) make this much easier to run.

* Small Cleanup Changes

Removed some unused imports