Intended Functionality
First off, I spotted an off-by-one in the skip function in the process of testing this, so that was fixed. Reacting to the message with the options provided by the bot now adds to the queue the selected track based on the user's reaction. There's some quality of life changes to be made (it doesn't check if it's in a voice channel on adding the track right now) and I'll get to those after lunch.