Feat/result picker (#4)
* Initial Commit

No attempt at the functionality quite yet, just the sketch of the idea in a block comment. Now to actually start implementing stuff.

* Embedded Message + Selector

The bot now sends an embedded message listing up to the first 5 search results for the given query, then reacts with emoji 1 through 5, to let the user select an item. Actually selecting an item is a job for another day, I'm clocking out.

* Reaction sensing

The bot now recognizes when a user reacts to it's embedded message, and, given the reaction they make, gathers the chosen url to pass to the ordinary player.

* Intended Functionality

First off, I spotted an off-by-one in the skip function in the process of testing this, so that was fixed. Reacting to the message with the options provided by the bot now adds to the queue the selected track based on the user's reaction. There's some quality of life changes to be made (it doesn't check if it's in a voice channel on adding the track right now) and I'll get to those after lunch.

* Code Cleanup

Gave the reaction subscriber and message update subscriber their own classes, in order to clean things up a bit. Also added some sass (read: messages) to certain cases, rather than just having the bot wordlessly do things. Up next is looking into deleting the embedded message after the user picks the item, but I might do that in another branch, as this one needs merging into develop.