Release Version 1.0.0 (#5)
* Fully Untested Attempt

Currently just pasting in the stuff from the example bot, going to sort out just what it's doing later, as I'm clocking out for the day.

* Additional Tinkering

Made some changes to the copied and pasted example bot code. As I expected, as time goes by, the code becomes more and more my own.

* Initial Commit

Develop is designed to be the most recent working version of the bot, which the feat/ branches branch off of. As the features reach a point where they can get added to the bot proper, they'll be merged into develop.

* Feat/normal youtube player (#1)

* Addition of Join Command

Without the join command, one would have to run the yt command twice, once to have the bot join you in a voice channel, and a second time to play the intended song. That simply won't do. (We have the yt command call the join command on not being in a voice channel anyway as before)

* Addition of Leave command

If there's a join command, there should be a leave command too, no?

* Inclusion of LeaveUser mapping

I wrote the actual code for the leave command, then forgot to map the command! Whoops.

* Feat/unique channel search (#2)

* Initial Attempt

The initial, untested attempt at the implementation of the bot's intended function: searching of a specific channel's video log. I've made it such that, if desired, one could use their own api key/channel id to use either their own resources or a youtube channel of their choosing, respectively.

* Search Implemented

After a great deal of toil in testing, the unique channel search works! Rather than using a direct filepath like I usually would when accesssing files, I found a solution that gets files out of a resources folder, provided it's in the right place. So, that's what is at play in fetching the properties. Going to implement that for my discord key as well, to (in theory) make this much easier to run.

* Small Cleanup Changes

Removed some unused imports

* Feat/logging and small todos (#3)

* Initial Commit

Dug through the project, found myself some updates to work on, and added TODO comments to make myself notice them.

* Added Logging (untested)

Added a file-based log for informational outputs, as well as the log for error-based outputs. Attempted to include slf4j to solve an odd error the discord api was throwing, but to little sucess. Will test further tomorrow, I'm off work now.

* Cleanup, Log Improvement, Properties Usage

First of all, implemented the use of a properties file for the discord token, a la how it's used for the youtube api key/channel id. Secondly, updated the log messages to make a little more sense, and thirdly, while updating the log messages, cleaned up any extraneous imports.

* Further Cleanup, Small Fixes

Cleaned up some more unused imports, as well as removed some unnecessary log messages. Additionally, the skip command originally skipped 1 less song than asked for, so that needed to be fixed. At this point, I'm recommending the branch be merged into dev

* Added attribution + small change

Made the message that the bot outputs on reacting to the track skip command make a bit more sense, and actually put my name on the bot. For real this time, ready for pull request, just had to double check I didn't have any outstanding small todos.

* Feat/result picker (#4)

* Initial Commit

No attempt at the functionality quite yet, just the sketch of the idea in a block comment. Now to actually start implementing stuff.

* Embedded Message + Selector

The bot now sends an embedded message listing up to the first 5 search results for the given query, then reacts with emoji 1 through 5, to let the user select an item. Actually selecting an item is a job for another day, I'm clocking out.

* Reaction sensing

The bot now recognizes when a user reacts to it's embedded message, and, given the reaction they make, gathers the chosen url to pass to the ordinary player.

* Intended Functionality

First off, I spotted an off-by-one in the skip function in the process of testing this, so that was fixed. Reacting to the message with the options provided by the bot now adds to the queue the selected track based on the user's reaction. There's some quality of life changes to be made (it doesn't check if it's in a voice channel on adding the track right now) and I'll get to those after lunch.

* Code Cleanup

Gave the reaction subscriber and message update subscriber their own classes, in order to clean things up a bit. Also added some sass (read: messages) to certain cases, rather than just having the bot wordlessly do things. Up next is looking into deleting the embedded message after the user picks the item, but I might do that in another branch, as this one needs merging into develop.

* Maven Fiddling

Trying to make it able to compile one runnable jar to release

* Maven Fiddling

A great deal of maven fiddling later, I've managed to make a jar file that I can release. Will sort that out as an official release via github.

* Prep for Release

Made some small changes to the project code to prepare for the 1.0.0 release. This isn't the end of the project, just a point at which the bot is "ready" to be out and about in the public.

* Update Readme

Forgot to update the readme to include the naptime command, that closes the bot, in case you ever need to shut it down for a little while.

* Update Readme

Update to the github readme to point out the releases folder, as that's rather important.