Feat/change items (#6)
* Initial Commit

Just the basic structure, and a small re-arrangement of items to make things cleaner. Going to work on the actual command after lunch.

* Prefix change command

Prefix change command is implemented, and protected such that only server managers or admins can use it. Need to push a bug fix to release to protect the naptime command, similarly

* Secured Shutdown

Secured the shutdown command to only allow admins and server managers to use it, and updated the release jar. Firmly suggest that this bug-fix version is used over the original.

* Search Length Change

Lets someone with the correct permissions change how many items that they want to show up in the search function. Locked similarly to the other bot-editing features.

* Update to remove version dependency

I noticed that the file paths for the logging would cause a version dependency for windows. In theory, this should allow the bot to run anywhere.