Next Release Version (#8)
* Search Length Change

Lets someone with the correct permissions change how many items that they want to show up in the search function. Locked similarly to the other bot-editing features.

* Update to remove version dependency

I noticed that the file paths for the logging would cause a version dependency for windows. In theory, this should allow the bot to run anywhere.

* Delete Button function (#7)

Added a reaction ❌ which the user can use to delete the message, in order to clean up the channel a bit.

* Merge branch 'master' into develop

* Update README.md

* Include Release Jar

Since we're merging dev into master, it needs to have the most recent version of the packaged jar.