Update for Release 1.2.0 (#10)
* Properties File Wizard (#9)

* Getting UI to Launch

Initial setup for the UI. Sure, all it does for now is show a screen that greets you, but it's only a few small steps to the actual intended functionality.

* Attempt at functionality (untested)

I haven't tested it quite yet, but in theory, you should be able to input the various items you need, and then hit the button. After pressing the button, it should generate a file (in the same directory) called discordbot.properties

* Post-Test Changes

This portion of the program works as intended now. Essentially, on opening the jar without a file named discordbot.properties, it'll open a little helpful wizard to make the required file.

* Include release jar

Include the most recent (soon to be released) build in the releases folder.