Untested JSON Backend
In theory, this should either create a properties json file if one doesn't exist, or, if the user has chosen a name for the bot that's not in the existing properties json, adds a bot's details to the properties. Up next is the choosing of the bot's name, which is going to be a javafx gui. That's a task for tomorrow, along with fixing all the calls to the former properties object.