Feat/improved resource file (#11)
* Initial Commit

The idea behind this feature branch is to replace the properties file with a JSON using google's simple json library. I'm also very tempted to make this a gradle project over maven, we shall see if I can pull that off.

* Conversion to Gradle

I'm fairly certain most of the conversion to a gradle project was on my end, unseen by you folks on github, but hey, now I've got much better control of the project. In short, replaced the pom.xml with a build.gradle file.

* End of Day Commit

Have to commit my code at the end of the day, just in case. Will start digging into actually replacing the properties file with a json tomorrow.

* Untested JSON Backend

In theory, this should either create a properties json file if one doesn't exist, or, if the user has chosen a name for the bot that's not in the existing properties json, adds a bot's details to the properties. Up next is the choosing of the bot's name, which is going to be a javafx gui. That's a task for tomorrow, along with fixing all the calls to the former properties object.

* (Untested) Attempt at name choosing

The initial attempt at picking a bot name. Had to use something other than javafx to have that dialog, which is probably for the best.

* Tested Name Input/JSON files

I'm fairly certain this closes out this branch, so after merging this into develop, the next step is new commands, first easter eggs, then digging into the youtube api for a "top ten" command, to get the top ten videos for the chosen channel key.