Tags Setting
Tags setting is implemented! That's all three of the twitch requests! hooray! All I need to do now is update the project file creator to have fields for, and the ability to write, the correct data into the properties file, then we should be ready to merge into dev and release! I'm real pumped about it!