Feat/twitch requests (#13)
* Initial structure

Structure of what twitch commands I want to happen. I should be able to use the v5 api for the title and game, and the new api for tags. Hopefully this helps me sort out OAuth whatnots so we can come back to the top ten youtube pull.

* Title and Game Setting

We can now set the title and game of the twitch channel provided on bot creation, provided the channel ID and the correct oauth key for said channel. We shall see if I can just continue to use the client ID for my twitch app for everyone, or if I need to have the user get one themself (like it's currently set up)

* Minor Tweaks

Just a few small comments to remove. Might add some more comments detailing what those ternary operators do, since ternary operators as a rule seem to be unclear to my eye.

* Idea behind Tags Setting

I detail the plan for the tags setter. Will work on this after lunch! (Step 1 is the embedded messages, step 2 is the sending of tags based on that embedded message)

* General idea of tags setter

The tags setter isn't quite done yet, but hey, I've got the general idea on this one! Wonderful! Shouldn't be too long 'til we finish this up!

* Lunchtime Commit

More structure for the tags changer. Lots to do, but I look forward to it. Coming back after lunch!

* Reactions

The bot provides a list of buttons to press, but none of them technically work yet. That's work for tomorrow, I'm out of time today.

* Arrow Functionality (Sorta)

The arrow functionality works as intended, but sometimes doesn't give buttons for the first embed on getting more than 10 tags. Will explore fixes.

* Tags Setting

Tags setting is implemented! That's all three of the twitch requests! hooray! All I need to do now is update the project file creator to have fields for, and the ability to write, the correct data into the properties file, then we should be ready to merge into dev and release! I'm real pumped about it!

* Ready for release!

Everything's running as intended, and I've updated the ReadMe to include the new things! Time to merge into dev, then merge that into master, since it's release time!